Moved userspace libc to posix container directory.
Issues:
- MM0 needs separate subset of libc and not link with libposix
- FS0 needs separate subset of libc and not link with libposix
- libposix need transforming into a full-blown uclibc for test0
- libc really is a pseudo-libc that needs to be replaced as above.
- Done some hacking to avoid conflicts between pseudo-libc and libposix headers.

Now system working as original, (e.g. execve etc. working)

+++ containers/posix/mm0/src/utcb.c
@@ -140,7 +140,7 @@ out:
 		if (IS_ERR(err = do_mmap(0, 0, task, slot,
 					 VMA_ANONYMOUS | VMA_PRIVATE |
 					 VMA_FIXED | VM_READ | VM_WRITE, 1))) {
-			printf("UTCB: mmapping failed with %d\n", err);
+			printf("UTCB: mmapping failed with %d\n", (int)err);
 			return (int)err;
 		}
 	}
